Cambios

+++ src/controllers/user.controller.ts
@@ -7,7 +7,7 @@ const userController = {
       const users = await User.find().populate('role');
       res.status(200).json(users);
     } catch (error) {
-      res.status(500).json({ error: error.message });
+      res.status(500).json({ error: 'ErrorgetAll' });
     }
   },
 
@@ -18,7 +18,7 @@ const userController = {
       await user.save();
       res.status(201).json({ message: 'User created successfully' });
     } catch (error) {
-      res.status(500).json({ error: error.message });
+      res.status(500).json({ error: 'ErrorCreate' });
     }
   },
 
@@ -31,7 +31,7 @@ const userController = {
       }
       res.status(200).json(user);
     } catch (error) {
-      res.status(500).json({ error: error.message });
+      res.status(500).json({ error: 'Errorgetbyid' });
     }
   },
 
@@ -50,7 +50,7 @@ const userController = {
       await user.save();
       res.status(200).json({ message: 'User updated successfully' });
     } catch (error) {
-      res.status(500).json({ error: error.message });
+      res.status(500).json({ error: 'Errorupdate' });
     }
   },
 
@@ -61,10 +61,14 @@ const userController = {
       if (!user) {
         return res.status(404).json({ error: 'User not found' });
       }
+      if (!('_id' in user)) {
+        throw new Error('Invalid user object');
+      }
       await user.delete();
       res.status(200).json({ message: 'User deleted successfully' });
     } catch (error) {
-      res.status(500).json({ error: error.message });
+      console.error(error); // registrar el error en la consola
+      res.status(500).json({ error: 'Error deleting user' });
     }
   }
 };
